removed async_reader.rs; move async impl to reader.rs; clean up
